tables analize
